tests/wpt/tests/dom/nodes/getElementsByClassName-17.htm

<!doctype html>
<html>
 <head>
  <title>document.getElementsByClassName(array): "b a"</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body class="a b">
  <div id="log"></div>
  <script>test(function() {
                 assert_array_equals(document.getElementsByClassName(["b a"]), [document.body]);
          })
 </script>
 </body>
</html>
